Reorganised HTTP demo packages.  Separating "long" and "short" protocol versions.
