+++ modules/core/src/test/scrib/demo/http/longvers/HttpClient.java
@@ -8,13 +8,13 @@ import org.scribble.net.session.SessionEndpoint;
 import org.scribble.net.session.SocketChannelEndpoint;
 import org.scribble.util.Caller;
 
-import demo.http.longvers.Http.Http.Http;
-import demo.http.longvers.Http.Http.channels.C.Http_C_1;
-import demo.http.longvers.Http.Http.channels.C.Http_C_4_Cases;
-import demo.http.longvers.Http.Http.channels.C.Http_C_5;
-import demo.http.longvers.Http.Http.channels.C.Http_C_5_Cases;
-import demo.http.longvers.Http.Http.channels.C.ioifaces.Branch_C_S_200__S_404.Branch_C_S_200__S_404_Enum;
-import demo.http.longvers.Http.Http.roles.C;
+import demo.http.longvers.HttpLong.Http.Http;
+import demo.http.longvers.HttpLong.Http.channels.C.Http_C_1;
+import demo.http.longvers.HttpLong.Http.channels.C.Http_C_4_Cases;
+import demo.http.longvers.HttpLong.Http.channels.C.Http_C_5;
+import demo.http.longvers.HttpLong.Http.channels.C.Http_C_5_Cases;
+import demo.http.longvers.HttpLong.Http.channels.C.ioifaces.Branch_C_S_200__S_404.Branch_C_S_200__S_404_Enum;
+import demo.http.longvers.HttpLong.Http.roles.C;
 import demo.http.longvers.message.Body;
 import demo.http.longvers.message.HttpMessageFormatter;
 import demo.http.longvers.message.client.Host;
