+++ modules/core/src/test/scrib/demo/http/longvers/HttpLong.scr
@@ -5,7 +5,7 @@
 //$ bin/scribblec.sh -ip modules/core/src/test/scrib/ -d modules/core/src/test/scrib/ modules/core/src/test/scrib/demo/http/longvers/Http.scr -api Http C -V
 //$ javadoc -cp modules/cli/target/classes/';'modules/core/target/classes';'modules/trace/target/classes';'modules/parser/target/classes';c:\Users\Raymond\.m2\repository\org\antlr\antlr-runtime\3.2\antlr-runtime-3.2.jar;'modules/validation/target/classes/';'modules/projection/target/classes/';C:\Users\Raymond\.m2\repository\org\codehaus\jackson\jackson-mapper-asl\1.9.9\jackson-mapper-asl-1.9.9.jar;C:\Users\Raymond\.m2\repository\org\codehaus\jackson\jackson-core-asl\1.9.9\jackson-core-asl-1.9.9.jar;'modules/core/src/test/scrib modules/core/src/test/scrib/demo/http/longvers/Http*.java -d test/doc/http/longvers
 
-module demo.http.longvers.Http;
+module demo.http.longvers.HttpLong;
 
 /*Connected to www.doc.ic.ac.uk.
 Escape character is '^]'.
