+++ modules/core/src/test/scrib/demo/http/longvers/message/Body.java
@@ -1,6 +1,6 @@
 package demo.http.longvers.message;
 
-import demo.http.longvers.Http.Http.Http;
+import demo.http.longvers.HttpLong.Http.Http;
 
 public class Body extends HttpMessage
 {
