+++ modules/core/src/test/scrib/demo/http/longvers/message/HttpMessage.java
@@ -3,7 +3,7 @@ package demo.http.longvers.message;
 import org.scribble.net.ScribMessage;
 import org.scribble.sesstype.name.Op;
 
-import demo.http.longvers.Http.Http.Http;
+import demo.http.longvers.HttpLong.Http.Http;
 
 // Unlike ScribMessage, HttpMessage is not actually "sent", but we use it as the base class since the socket API takes ScribMessages
 public abstract class HttpMessage extends ScribMessage
