+++ modules/core/src/test/scrib/demo/http/longvers/message/server/LastModified.java
@@ -1,6 +1,6 @@
 package demo.http.longvers.message.server;
 
-import demo.http.longvers.Http.Http.Http;
+import demo.http.longvers.HttpLong.Http.Http;
 import demo.http.longvers.message.HeaderField;
 
 public class LastModified extends HeaderField
