+++ modules/core/src/test/scrib/demo/http/longvers/message/server/_404.java
@@ -1,6 +1,6 @@
 package demo.http.longvers.message.server;
 
-import demo.http.longvers.Http.Http.Http;
+import demo.http.longvers.HttpLong.Http.Http;
 
 public class _404 extends StatusCode
 {
