+++ modules/core/src/test/scrib/demo/http/shortvers/HttpShort.scr
@@ -1,0 +1,39 @@
+//Raymond@HZHL3 ~/code/scribble-java/scribble-java
+//$ bin/scribblec.sh -ip modules/core/src/test/scrib/ -d modules/core/src/test/scrib/ modules/core/src/test/scrib/demo/http/shortvers/HttpShort.scr -api Http C -V
+//$ javadoc -cp modules/cli/target/classes/';'modules/core/target/classes';'modules/trace/target/classes';'modules/parser/target/classes';c:\Users\Raymond\.m2\repository\org\antlr\antlr-runtime\3.2\antlr-runtime-3.2.jar;'modules/validation/target/classes/';'modules/projection/target/classes/';C:\Users\Raymond\.m2\repository\org\codehaus\jackson\jackson-mapper-asl\1.9.9\jackson-mapper-asl-1.9.9.jar;C:\Users\Raymond\.m2\repository\org\codehaus\jackson\jackson-core-asl\1.9.9\jackson-core-asl-1.9.9.jar;'modules/core/src/test/scrib modules/core/src/test/scrib/demo/http/shortvers/Http*.java -d test/doc/http/shortvers
+
+module demo.http.shortvers.HttpShort;
+
+/*Connected to www.doc.ic.ac.uk.
+Escape character is '^]'.
+GET /~rhu/ HTTP/1.1
+host: foo.bar.com
+
+HTTP/1.1 200 OK
+Date: Mon, 13 Jun 2016 19:42:34 GMT
+Server: Apache
+Strict-Transport-Security: max-age=31536000; preload; includeSubDomains
+Strict-Transport-Security: max-age=31536000; preload; includeSubDomains  // BUG? (Apache configuration)
+Last-Modified: Thu, 14 Apr 2016 12:46:24 GMT
+ETag: "74a-53071482f6e0f"
+Accept-Ranges: bytes
+Content-Length: 1866
+Vary: Accept-Encoding
+Content-Type: text/html
+Via: 1.1 www.doc.ic.ac.uk*/
+
+
+sig <java> "demo.http.shortvers.message.client.Request"
+		from "demo/http/shortvers/message/Request.java"
+		as REQUEST;
+
+sig <java> "demo.http.shortvers.message.server.Response"
+		from "demo/http/shortvers/message/Response.java"
+		as RESPONSE;
+
+
+global protocol Http(role C, role S)
+{
+	REQUEST from C to S;
+	RESPONSE from S to C;
+}
